requirements.txt
torch
transformers>=4.55.0
trl>=0.20.0
peft>=0.17.0
wandb==0.16.0
nltk
rouge_score
pyyaml
numpy
tqdm